requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
marshmallow==3.14.1
python-dotenv==0.19.2
SQLAlchemy==1.3.13
psycopg2-binary==2.9.5
twilio==7.13.0
requests==2.28.1
pytest==6.2.4
PyPDF2==2.4.1
regex==2022.7.25
redis==4.3.4
boto3==1.24.59
PyJWT==2.4.0
aws-secretsmanager-caching==1.1.1.5
